Fix missing event page labels (#557)
Fix the issue where sometimes event page labels were missing in ExUI after the case definition was generated by the CCD config generator. 

Assign the page label to all fields in the event page because the fields are sorted when written to the JSON but CCD uses the page label in the first field of the serialised case definition, (not the order of the fields in the Java config).